requirements.txt
requests == 2.28.0
python-dotenv == 0.20.0
pandas==1.4.3